refactor: 게시글 작성 mutation options 분리
- type도 api/ 디렉토리 안으로 이동
