+++ src/api/types.ts
@@ -1,3 +1,6 @@
+import { ApiAnimalType } from "@/types/animal";
+import { ApiEmotion } from "@/types/Emotion";
+
 export interface ILoginCode {
   code: string;
 }
@@ -23,4 +26,51 @@ export interface IUserRequest {
   nickname: string;
   animalType: string;
   profileImage: File | null;
+}
+
+export interface ICreatePost {
+  images: File[];
+  content: string;
+  emotion: ApiEmotion;
+}
+
+export interface IPostDetailData {
+  id: number;
+  userId: number;
+  nickname: string;
+  profileImageUrl: string;
+  transformedContent: string;
+  emotion: ApiEmotion;
+  postType: ApiAnimalType;
+  imageUrls: string[];
+  commentCount: number;
+  likeCount: number;
+  liked: boolean;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface IPostSummaryData {
+  id: number;
+  userId: number;
+  nickname: string;
+  profileImageUrl: string | null;
+  transformedContent: string;
+  emotion: ApiEmotion;
+  postType: ApiAnimalType;
+  thumbnailUrl: string | null;
+  commentCount: number;
+  likeCount: number;
+  liked: boolean;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface IPostSummaryDataPagination {
+  currentPage: number;
+  last: boolean;
+  pageSize: number;
+  totalElements: number;
+  totalPages: number;
+  content: IPostSummaryData[];
 }
